% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotpuntos.R
\name{plotpuntos}
\alias{plotpuntos}
\title{Plot puntos: figura de nube de puntos}
\usage{
plotpuntos(x, y)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vectos}
}
\value{
a plot of point
}
\description{
Plot puntos: figura de nube de puntos
}
\examples{
plotpuntos(c(2,3,4),c(3,5,6))
}
