% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resumenstats.R
\name{resumen}
\alias{resumen}
\title{Resumen exploratorio de datos (Analisi)}
\usage{
resumen(x)
}
\arguments{
\item{x}{nimeric vector}
}
\value{
mean and sd from vector x
}
\description{
Resumen exploratorio de datos (Analisi)
}
\examples{
resumen(c(1,2,3,4,8))
}
