% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotpuntos.R
\name{plotpuntos}
\alias{plotpuntos}
\title{plotpuntos: figura de nube de puntos}
\usage{
plotpuntos(x, y)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector}
}
\value{
a simple plot of points
}
\description{
plotpuntos: figura de nube de puntos
}
\examples{
plotpuntos(c(2,3,4), c(3,4,6))

}
