% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resumen.R
\name{resumen}
\alias{resumen}
\title{Resumen: Analsis exploratorio de datos}
\usage{
resumen(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
mean and sd from vector x
}
\description{
Resumen: Analsis exploratorio de datos
}
\examples{
resumen(c(1,2,3,4,8))

}
