% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resume.R
\name{resumen}
\alias{resumen}
\title{Resumen: Análisis exploratorio de datos}
\usage{
resumen(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
mena and sd from vector x
}
\description{
Resumen: Análisis exploratorio de datos
}
\examples{
resumen(c(1, 2, 3, 4, 8))

}
